Projects/Agentic_AI_Based_Task_Management_System_backend/pyproject.toml
[project]
name = "agentic-ai-based-task-management-system-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "chainlit>=2.9.4",
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "google-genai>=1.57.0",
    "httpx>=0.28.1",
    "openai-agents>=0.6.5",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
    "streamlit>=1.52.2",
    "uvicorn>=0.40.0",
]
